Refactor: add shift module and moved methods into it, cleaned up other modules and tests
