+++ spec/shifts_spec.rb
@@ -1,0 +1,1 @@
+require './spec_helper'
