+++ spec/spec_helper.rb
@@ -6,3 +6,4 @@ require './lib/enigma'
 require './lib/encrypt'
 require './lib/decrypt'
 require './lib/generator'
+require './lib/shifts'
